requirements.txt
streamlit==1.34.0
openai==1.25.1
gspread==6.0.2
google-auth==2.29.0
python-dotenv==1.0.1
pandas==2.2.2 